coffeShop/requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
pillow==10.4.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.9